refactor: singular package name
